+++ relay-server/src/metrics_extraction/mod.rs
@@ -1,2 +1,3 @@
 pub mod sessions;
 pub mod transactions;
+mod utils;