+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Aggregate outcomes for dynamic sampling, invalid project ID, and rate limits. ([#1134](https://github.com/getsentry/relay/pull/1134))
 - Extract session metrics from aggregate sessions. ([#1140](https://github.com/getsentry/relay/pull/1140))
 - Prefix names of extracted metrics by `sentry.sessions.` or `sentry.transactions.`. ([#1147](https://github.com/getsentry/relay/pull/1147))
+- Extract transaction duration as metric. ([#1148](https://github.com/getsentry/relay/pull/1148))
 
 ## 21.11.0